commit 8

+++ src/entities/Categories.java
@@ -1,10 +1,10 @@
 package entities;
 
 public class Categories {
-    private Long id;
+    private Integer id;
     private String categorie;
 
-    public Categories(Long id, String categorie) {
+    public Categories(Integer id, String categorie) {
         this.id = id;
         this.categorie = categorie;
     }
@@ -13,11 +13,11 @@ public class Categories {
 
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
